+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - custom operator `..` should not cover `../`
 - optimize current dir read from env first.
 - fix relative path for pty cmd.
-
+- allow relative path based on `SCRIPT` env for `include` and `import`, and `config` file
 
 ## [0.12.3]